+++ src/tx-decorator/convert-buy-pool.js
@@ -1,0 +1,3 @@
+import decorateBuySwapPoolTxParams from './convert-sell-pool.js';
+
+export default decorateBuySwapPoolTxParams;