+++ src/tx-decorator/convert-sell-all-pool.js
@@ -1,0 +1,10 @@
+/**
+ * @param {TxParams} params
+ * @return {TxParams}
+ */
+export default function decorateSellAllSwapPoolTxParams(params) {
+    // force gasCoin to be same as coin to sell
+    params.gasCoin = params.data.coins[0];
+
+    return params;
+}