+++ src/tx-decorator/convert-sell-all.js
@@ -1,3 +1,10 @@
-import decorateSellAllTxParams from './convert-sell.js';
+/**
+ * @param {TxParams} params
+ * @return {TxParams}
+ */
+export default function decorateSellTxParams(params) {
+    // force gasCoin to be same as coinToSell
+    params.gasCoin = params.data.coinToSell;
 
-export default decorateSellAllTxParams;
+    return params;
+}