+++ src/tx-decorator/convert-sell-pool.js
@@ -2,7 +2,7 @@
  * @param {TxParams} params
  * @return {TxParams}
  */
-export default function decorateSwapPoolTxParams(params) {
+export default function decorateSellSwapPoolTxParams(params) {
     if (typeof params.gasCoin === 'undefined') {
         params.gasCoin = params.data.coins[0];
     }